fix: use token from process.env
